city_pulse/requirements.txt
google-cloud-bigquery==3.25.0
google-cloud-logging==3.11.0
google-adk==0.1.0
google-genai==0.2.0
pandas==2.2.0
geopandas==0.14.3
folium==0.15.1
plotly==5.18.0
python-dotenv==1.0.0
shapely==2.0.2
matplotlib==3.8.2
seaborn==0.13.1

